refactor(native_cc, main): portable basename helper + extract clap derives
Two #310 follow-up fixes plus a structural split:

1. native_cc.rs::value_starts_with_known_wrapper used Path::new(s).file_stem()
   which only honours the host OS's separator — so on Linux/macOS, a
   Windows-style `C:\Tools\sccache.exe ...` value was NOT recognised as
   a known wrapper, breaking the no-double-wrap invariant. Factored a
   `portable_basename_stem` helper that splits on both `/` and `\`
   regardless of host. Surfaced by the existing
   `detects_wrapper_paths_with_extension` test, which now passes on
   Linux/macOS CI.

2. native_cc_tests.rs::inject_wraps_target_specific_when_user_set_them
   called wrap_compiler_env with 3 args; the function takes 4 since
   #310 added synthesize_default. Pass `true` for parity with the other
   user-set-env tests in the file. Pre-existed on main but was masked
   on Windows by the test's `#[cfg(not(target_os = "windows"))]` gate.

3. main.rs grew to 1503 LOC which hard-blocks the loc_guard hook on
   every edit. Moved all clap derives (Cli, Commands, *Subcommand,
   GcCargoArgs, GcSweepArgs, TrimProfileArg, ZccacheSourceArg, the
   SOLDR_BUILTIN_VERBS const) into a new `cli_args` submodule, leaving
   main.rs at 837 LOC with headroom. cli_args.rs is 682 LOC — also
   under the warn threshold. Re-exports preserve the dispatch path so
   `run_cli` is byte-identical.
